+++ ariles/tests/api_v2/regression_test_240.cpp
@@ -1,0 +1,62 @@
+/**
+    @file
+    @author  Alexander Sherikov
+
+    @copyright 2017-2026 Alexander Sherikov, Licensed under the Apache License, Version 2.0.
+    (see @ref LICENSE or http://www.apache.org/licenses/LICENSE-2.0)
+
+    @brief
+*/
+
+#include "utility.h"
+
+#include "all_enabled_visitors.h"
+#include "all_enabled_adapters.h"
+
+#define ARILES2_DEFAULT_VISITORS                                                                                       \
+    ARILES2_VISITOR(count)                                                                                             \
+    ARILES2_VISITOR(count_missing)                                                                                     \
+    ARILES2_VISITOR(finalize)                                                                                          \
+    ARILES2_VISITOR(prewrite)                                                                                          \
+    ARILES2_VISITOR(defaults)                                                                                          \
+    ARILES2_VISITOR(read)                                                                                              \
+    ARILES2_VISITOR(write)                                                                                             \
+    ARILES2_VISITOR(compare)
+
+#include <ariles2/visitors/compare.h>
+#include <ariles2/ariles.h>
+
+
+// ===============================================================
+// TYPES
+// ===============================================================
+
+#include "types/partial.h"
+
+
+// ===============================================================
+// FIXTURES
+// ===============================================================
+
+#include "fixtures/initializers.h"
+#include "fixtures/025_comparison_partial.h"
+
+
+// ===============================================================
+// TESTS
+// ===============================================================
+
+
+
+#define ARILES_TESTS(VISITOR_ID, NAMESPACE, INITIALIZER)                                                               \
+    BOOST_FIXTURE_TEST_CASE(                                                                                           \
+            PartialEquivalenceFixture##_##VISITOR_ID##_##INITIALIZER,                                                  \
+            ariles_tests::PartialEquivalenceFixture<ariles_tests::initializers::INITIALIZER>)                          \
+    {                                                                                                                  \
+        test<ariles_tests::Partial, ariles_tests::Complete, ariles2::NAMESPACE>();                                     \
+    }
+
+
+// wont work
+#define ARILES2_TESTS_DISABLE_msgpack_compact
+#include "instantiate.h"